Cast warm-started scan/tilt to the plan dtype on checkpoint resume
probe.data and obj.data were already cast to the run's dtype when re-used
from a loaded checkpoint (init_state), but scan and tilt weren't -- state.to_xp()
only moves arrays to the right backend (xp.asarray), it doesn't touch dtype.
Resuming a run whose checkpoint was saved at a different float precision than
the current plan's `dtype` therefore left `scan`/`tilt` at the checkpoint's
original dtype, surfacing downstream as a jax.lax.scan dtype-mismatch
TypeError partway into the resumed run.

Found independently on SuperCloud (four warm-started jobs crashing
identically at iteration 5); mirrored here so it isn't lost when the branch
there gets synced back.

+++ phaser/execute.py
@@ -306,7 +306,7 @@ def initialize_reconstruction(
 
     if init_state.scan is not None and plan.init.scan is None:
         logging.info("Re-using scan from initial state...")
-        scan = init_state.scan
+        scan = init_state.scan.astype(dtype)
     else:
         logging.info("Initializing scan...")
         scan = pane.from_data(scan_hook, ScanHook)(  # type: ignore
@@ -315,7 +315,7 @@ def initialize_reconstruction(
 
     if init_state.tilt is not None and plan.init.tilt is None:
         logging.info("Re-using tilt from initial state...")
-        tilt = init_state.tilt
+        tilt = init_state.tilt.astype(dtype)
     elif tilt_hook is not None:
         logging.info("Initializing tilt...")
         tilt = pane.from_data(tilt_hook, TiltHook)(  # type: ignore
